Added proper logging
Add a hibernate logging facility. Logging is tricky in hibernate because
there are times when neither the syslog nor the file system is
unavailable. Instead, we fallocate more disk files to hold the logs
while suspending, and the logs while resuming. Any kernel (failed
suspend, bootstrap, or resumed original) can replay these log files to
collect what happened during the dark period of suspend or resume.

To keep things in a comprehensible order, there are also times when we
may want to just buffer log messages and keep them in memory. So the
overall log flow looks like this:
 * Hibernate setup portion goes straight to syslog.
 * Just before taking the snapshot, switch to suspend_log diskfile.
 * Attempt to finish, flush to that disk file, and power off.
 * On resume, a couple of logs to go the syslog (to be discarded), but
   fairly early we switch to logging into the resume_log diskfile.
 * Load the hibernate image, flush the logs to the resume file, and
   jump to the new kernel.
 * The resumed system immediately switches to buffering logs in-memory.
 * It finishes resume, and restarts tasks (including the syslogger).
 * It then replays the hibernate log, followed by the resume log, into
   the syslog.
 * Finally, logs that had been buffered in memory are flushed to syslog,
   and normal logging directly to syslog resumes.

If at any point there's a failure and the process is aborted, we can
still replay suspend logs, and possibly resume logs (depending on if we
know we got that far).

One challenge is that if we don't open the disk with O_DIRECT, then
previous reads will be cached, and we'll read a stale suspend log. We
can test this by running hiberman hibernate --dry-run, hiberman
hibernate, then hiberman resume. On resume, we'll see the dry run logs
if O_DIRECT was not specified.