Add hibernate image hint cookie
Early init needs a hint to know whether to mount stateful via
dm-snapshot (to keep the disk RO), or directly as an RW mount. Utilize
free space in the sector that stores the GPT header. This structure is
significantly smaller than even a 512 byte sector, is always likely
to consume a whole sector, is highly likely to stay in the same place,
and is unlikely to be tampered with during normal operation. As such,
it's an ideal candidate for this token of data that needs to live
outside an RW file system.

We add a check for the GPT partition magic as well to make sure we're
writing in the correct place. This change also adds a debugging command
that can get and set the cookie. With this, we'll be able to modify
chromeos_startup to divert to the dm-snapshot mode on resume.