Make resume work with dm-snapshot
After adding the scaffolding to setup dm-snapshot on resume boots, the
logic to find the stateful partition and underlying block device broke.
Use the "rootdev" app to figure out the underlying block device, then
build the partition on top of that. This hardcodes partition 1 as the
stateful partition, though that's very unlikely to change.