Add symmetric image encryption
Add the ability to encrypt the hibernate image with a symmetric key. For
now, we're just putting the key and IV directly in the hibernate
metadata, which will obviously change. Also add --unencrypted flags to
the command line. On hibernate this disables encryption when creating
the hibernate image. On resume it allows an unencrypted image to be
resumed, which would otherwise be rejected for security reasons.