Refactor metadata into public/private portions
Start going through more of the real motions with metadata. Split out
the unencrypted portion from the encrypted portion, and encrypt/decrypt
the encrypted portion. Right now we just use a fixed key for encryption,
but it serves as a more ready entry point that the asymmetric crypto can
plug into.

Split the process of loading the metadata from disk into two
phases: public and private. When the asymmetric crypto comes online
we'll need to load the public portion to get the ephemeral DH keys to
decrypt the private portion. So it has to be split into two phases then
anyway.