Add disabled preloader
This change adds a class that can preload the image before the
ImageMover pumps it through decryption and into the kernel. This is
supposed to be useful so we can load the hibernate image from disk while
the user it typing in their password.

Unfortunately I didn't realize the kernel attempts to preallocate the
entire hibernate image once it gets its header data. It obviously runs
out of space trying to this, since it wants ~50% of memory, hiberman
already has ~50% of memory, and the resume kernel+chrome are already
sitting on some memory.

Check this in but leave it disabled for now. I'm going to try a scheme
where we preload the hibernate header early (so the kernel gets its
chunk), and then use the remaining memory for preloading. We'll see how
far that gets.