Add ImageSplitter to make preloading work
The preloader had a problem in that there seemed to be plenty of memory
for it to load the whole image. But after feeding a few pages into the
kernel, the kernel suddenly exploded because it was out of memory. This
turns out to be because the kernel attempts to allocate the entire space
it will need for the hibernate image after receiving its header and page
list.

The header and page list are pretty boring structures. There's no PII in
there, and they're simple enough that the kernel can safely validate
them even as untrusted content. So introduce an ImageSplitter that reads
the hibernate header on its way down and diverts the header and pagelist
to a separate (unencrypted) file. We can then load this file immediately
at boot and feed it to the kernel, so the kernel can commence with its
giant allocation. After that happens, the preloading process can proceed
in a more predictable fashion, loading pages and watching for low
memory.

One thing that's still needed is to hash the header pages and save a
hash in the encrypted portion of the metadata, so we can detect
tampering with this portion of the image.