Add a D-bus server
Add a D-bus server, which spawns a new thread that simply waits for
someone to call its SetSeedMaterial function. This will be handed to us
by cryptohome and contains our secret material which will be used to
construct the hibernate public/private keypair.

Note that a piece of this is still missing: we don't yet wait on a
non-resume to accept the key from cryptohome and save it for a future
hibernate. That's up next.