Save and verify the header pages hash
The header and pagemap pages are being saved unencrypted. They contain
no PII, so this is fine, but they also need to be protected against
offline tampering.

Hash the header pages on the way down and save the result in the private
metadata. On resume, validate that the hash still matches once the
private metadata is available.

Note: I'd like to do this before moving the main image contents rather
than at the very end. Unfortunately I've (maybe) handed a mutable
reference to the joiner to the Preloader, so I can't touch the joiner to
ask for the hash. Instead I need to wait until the pipeline is finished
(ie the mover is done) to reclaim my joiner reference. This should be
ok, but I wonder if there's a better way.