Poison and clear metadata at resume
Clear out the metadata private area when we begin to attempt resume in
earnest. This mitigates against attacks that compromise the hibernate
key but only have disk read privileges by forgetting the symmetric key
used to encrypt the hibernate data. It also prevents against accidental
double resumes. As an additional breadcrumb (and since we're writing the
metadata anyway), split the RESUME flag into two: RESUME_STARTED to
indicate we wiped the private data, and RESUME_LAUNCHED to indicate we
actually tried to jump into this image.