Add CryptoReader for better crypto symmetry
I was hoping to be able to get away with just using CryptoWriter, and
inserting it at different points in the pipeline for image writes
and reads (just on the write side of the ImageMover for both).
Unfortunately this gives me no opportunity in the pipeline to operate on
the unencrypted data when resuming.

It turns out to coerce the kernel into doing its big image allocation, I
have to write all the meta pages, plus one byte of data. I can squirrel
away that first byte into metadata so I can push it before decryption,
but unfortunately there's no place for me to verify that byte in the
encrypted stream.

Adding a CryptoReader lets us put the decryption on the read side of the
ImageMover, which makes it easier to peek at an unencrypted byte before
letting the pump run continuously.