Fix overdecryption in CryptoReader
The read() function in CryptoReader is complicated because it's trying
to avoid copying large buffers unnecessarily. So it decrypts all except
the last block directly into the caller's buffer, then decrypts the last
block into an extra buffer, and loops back to copy that into the
caller's buffer if needed.

The bounds on that last block of decrypted data were wrong. If the
caller requested less than the internal buffer size, we were
accidentally decrypting not only the last block, but also the stale data
at the end of the internal buffer. Fix that boundary to get resume
working properly even when the ImageMover moves only a page at a time.