Push the first data byte into resume early
The kernel's implementation is annoying in that it will only do its
giant allocation once its received the first byte of non-metadata and is
looking for a place to put it. So if we only push the metadata (but no
data), the kernel still doesn't do its giant allocation, and then our
userspace preloader takes too much. Store the first byte of data
unencrypted in the metadata so we can feed it into the kernel early.
This "first byte" represents the first byte of the page of RAM with the
lowest physical address. This page is unlikely to be availble for
general user data, so an attacker is going to have a hard and slow time
using this one byte to exfiltrate data.