Refactor snapshot functions into SnapshotDevice struct
Factor out the functions dealing with the snapshot device, and create a
new SnapshotDevice structure to encapsulate those methods. Refactor the
methods themselves into using a common helper function to actually do
the ioctl and evaluate the result. This seems to result in better
separation of code, and a lot less code. Yay