Add documentation
Add a bunch of doc-u-comments to help describe the modules, functions,
constants, etc. Also add a big README.md to hopefully prime reviewers on
how hibernate works and how this utility is structured. Finally, add
Alexis as an OWNER.

Ideally this would be a total no-op comment-only change, but a couple of
random odds and ends did sneak in. Hopefully harmless.