Move File into HiberlogOut enum
I tried to figure this out earlier, but could not figure out how to
borrow the File from inside a custom enum. I get now that you have to
take a mutable reference on the enum, which also gives you permission to
mutate the thing inside the enum. This allows me to simplify the
redirect_log() API.