Use serde crate to do metadata serialization
My repr(C) apparently runs the risk of being unsound due to padding
bytes. Use serde instead as a blessed serialization mechanism.
Unfortunately serde doesn't support arrays of > 32 elements, so I can no
longer embed the encrypted private metadata blob in the public metadata
structure. So instead, write the private metadata separately at a known
offset in the metadata file. The encryption is also a bit of a pain,
since the output buffer must always be a block bigger than the input
buffer, and when encrypting a fixed size, we always round up to the next
block, even if the input is already block aligned.