Use RAII for unfreezing userspace in SnapshotDevice
Have freeze_userspace() return a wrapped version of the SnapshotDevice
that calls unfreeze_userspace() when dropped. It's a wrapper because
both the rest of the suspend/resume process and the
FrozenUserspaceTicket need to mutate the SnapshotDevice. So require that
the rest of the suspend/resume process keep the ticket, and get a
mutable reference to the snapshot device through it.