Use OnceCell in the Hiberlog singleton
Convert the Hiberlog singleton instance to use OnceCell. This required
converting the dyn Write trait object into a dyn Write + Send.
Unfortunately that dragged in MmapBuffer, which contains a raw pointer,
which is not Send. Safely add the Send trait to MmapBuffer since we can
guarantee this is never a TLS pointer.