Use OCB encryption for hibernate metadata
In order to get both encryption and authentication, use AES in OCB
mode so we can detect tampered-with metadata.

I had to make the IV size equal to the default IV size, because the Rust
openssl library inappropriately makes a GCM-specific call
(EVP_CTRL_GCM_SET_IVLEN) to set the IV length.