Correctly preallocate hiberfile when KASAN is enabled
I'm running a kernel with KASAN enabled. KASAN steals 1/8 of memory for
its reporting and makes it look invisible. But that memory still has to
be saved and restored as part of hibernate. So our maximum hibernate
image size calculation ended up being too small, and we failed to
hibernate because we hit the end of the preallocated file. For the very
rare case where KASAN is enabled, overprovision the hibernate image to
account for that extra 1/8th of memory.

Also check in a few of the additional context lines that helped me
figure out exactly what was going on.