Use OCB mode encryption for hibernate image data
Switch from aes_128_cbc, which has no protection against tampering, to
aes_128_ocb, which includes an authentication tag that will not match if
the ciphertext has been modified. The tag is added into the private
metadata portion.

Getting the lifetimes right such that we could read and write the tag in
CryptoReader/CryptoWriter required a little rework of the
read_image/write_image pipeline. In the old pipeline, we "maybe" created
a CryptoWriter, then immediately handed ownership to a &mut dyn Write
object, and inserted that into the pipeline. That meant we had no
ownership over it anymore to be able to call get_tag(), and only the
Write trait to work with in the objects that did own it. To fix this,
always instantiate and insert the CryptoReader/CryptoWriter structures
in the pipeline, but give them an unencrypted passthrough mode to handle
the unencrypted case. With this, the compiler is able to reason about
lifetimes better and understand that once the mover is dropped, the
CryptReader/CryptWriter is available for mutable use again.